src/requirements.txt
altair==5.5.0
auto_mix_prep==0.2.0
langchain==0.3.15
langchain_chroma==0.2.0
langchain_core==0.3.31
langchain_google_genai==2.0.9
pandas==2.2.3
protobuf==5.29.3
python-dotenv==1.0.1
regex==2024.11.6
